implemented langchain adapter
implemented langchain adapter based on existing langchain code, made sure no langchain involved in core rag, added unittest and integration test. made sure only unittest will be tested in ci
