+++ app/services/rag/adapters/__init__.py
@@ -1,0 +1,3 @@
+from app.services.rag.adapters.langchain_adapter import LangChainRAGAdapter
+
+__all__ = ["LangChainRAGAdapter"]
