+++ app/services/rag/generator/__init__.py
@@ -1,0 +1,3 @@
+from app.services.rag.generator.chatgpt_generator import ChatGPTGenerator
+
+__all__ = ["ChatGPTGenerator"]
