+++ docker-compose.yml
@@ -30,7 +30,8 @@ services:
     volumes:
       - .:/app
     environment:
-      - DATABASE_URL=postgresql+asyncpg://rag_user:rag_password@postgres:5432/rag_vectordb
+      - DATABASE_URL=postgresql://rag_user:rag_password@postgres:5432/rag_vectordb
+      - ASYNC_DATABASE_URL=postgresql+asyncpg://rag_user:rag_password@postgres:5432/rag_vectordb
 
   # GPU 工作节点 (给带 Nvidia 显卡的组员)
   worker-gpu:
@@ -47,7 +48,8 @@ services:
     volumes:
       - .:/app
     environment:
-      - DATABASE_URL=postgresql+asyncpg://rag_user:rag_password@postgres:5432/rag_vectordb
+      - DATABASE_URL=postgresql://rag_user:rag_password@postgres:5432/rag_vectordb
+      - ASYNC_DATABASE_URL=postgresql+asyncpg://rag_user:rag_password@postgres:5432/rag_vectordb
     deploy:
       resources:
         reservations:
